Tested working with db class file and object call
